fix: backend

+++ backend/internal/http/requests/comment_request.go
@@ -2,12 +2,12 @@ package requests
 
 type CommentCreateRequest struct {
 	Content string `json:"content" validate:"required,min=1,max=1000"`
-	UserID  uint   `json:"user_id" validate:"required,gt=0"`
-	PostID  uint   `json:"post_id" validate:"required,gt=0"`
+	UserID  uint64 `json:"user_id" validate:"required,gt=0"`
+	PostID  uint64 `json:"post_id" validate:"required,gt=0"`
 }
 
 type CommentUpdateRequest struct {
 	Content string `json:"content" validate:"required,min=1,max=1000"`
-	UserID  uint   `json:"user_id" validate:"required,gt=0"`
-	PostID  uint   `json:"post_id" validate:"required,gt=0"`
+	UserID  uint64 `json:"user_id" validate:"required,gt=0"`
+	PostID  uint64 `json:"post_id" validate:"required,gt=0"`
 }
