+++ backend/internal/http/requests/media_association_request.go
@@ -1,7 +1,7 @@
 package requests
 
 type MediaAssociationInput struct {
-	MediableID   uint   `json:"mediable_id" binding:"required"`
+	MediableID   uint64 `json:"mediable_id" binding:"required"`
 	MediableType string `json:"mediable_type" binding:"required"`
 	Group        string `json:"group" binding:"required"`
 }
@@ -11,7 +11,7 @@ type MediaCreateRequest struct {
 	URL    string `json:"url" validate:"required,url"`
 	Type   string `json:"type" validate:"required"`
 	Size   int64  `json:"size" validate:"required,gt=0"`
-	UserID uint   `json:"user_id" validate:"required,gt=0"`
+	UserID uint64 `json:"user_id" validate:"required,gt=0"`
 }
 
 type MediaUpdateRequest struct {
@@ -19,5 +19,5 @@ type MediaUpdateRequest struct {
 	URL    string `json:"url" validate:"required,url"`
 	Type   string `json:"type" validate:"required"`
 	Size   int64  `json:"size" validate:"required,gt=0"`
-	UserID uint   `json:"user_id" validate:"required,gt=0"`
+	UserID uint64 `json:"user_id" validate:"required,gt=0"`
 }
